All changes so Far:
implemented classes in renderer interface
made agent and projectile child classes of entity class so that they could be drawn in the renderer interface
moved code from main to raylib renderer
elle worked on raylib rendererer mostly
highlighted aspects in raylib renderer where we will need to call the functions that spawn entities and where we will need to draw said entities. 
That being said, all code is untested because the project does not compile on either mine or elle's computer. You will need to tell us if code works or not.